Change writer functions back to being static.
