+++ src/utils/index.ts
@@ -19,4 +19,3 @@ export * from "./StringUtils";
 export * from "./tsconfig";
 export * from "./TypeGuards";
 export * from "./TypeScriptVersionChecker";
-export * from "./WriterFunctions";
